refactor: move sanity checks into a single file
